Handling empty response when deleting
When working with deleting phone numbers we got a MultiJson decode error
and needed to better handle the case where the body was an empty string.

We also began the test by disallowing external connections to provide
a clearer error message when an external call is being made and it needs
to be faked out.