Handle 204 response from resource#delete
Net::HTTP returns a nil body when it receives a 204 response, which
caused the client to raise a method not found on nil for empty?.

FakeWeb doesn't allow nil response bodies, so we had to use more
intrusive stubbing. There is a pull request to fix this bug on FakeWeb
[https://github.com/chrisk/fakeweb/pull/32], so this test should be
updated upon FakeWeb being updated.