Updates rspec to v3. Removes Gemfile gems from gemspec.
The rspec syntax for the test was updated recently, but rspec itself wasn't bumped, so old installs would fail the tests.
Also, rspec, fakeweb and rack were all declared in both the Gemfile test group and as development dependencies in the gemspec. Whilst I'd rather have them as development dependencies, I noticed a commit from @karlfreeman that moved them to the Gemfile, so I will stick with this for the project. I have removed them again from the gemspec for consistency.