Updates the config block to work across more classes.
It didn't make sense for the Twilio::REST::Client to look after configuration, as both the Twilio::Util::Capability and the Twilio::Util::RequestValidator both use one or both of the credentials too. So, I moved the configure method to the Twilio module directly, the Configuration class to the util/ directory and updated the classes that take account_sid and auth_token as an initializer to use the config if it is available.