Stops modifying NextGenListResource @path ivar.
I'm not sure if this is necessarily the best way to do this right now, but it mirrors the behaviour in `Twilio::REST::Client` so is at least safe. The issue is that the path variable passed from the `NextGenListResource` to the `TaskRouterClient#get` method was being modified which altered it in the class. This lead to parameters being piled onto the end of an ever growing `@path` variable as more calls were made to `list`.