Allow repeated parameter names for lookup requests
The API tester at https://www.twilio.com/lookup has a repeated
"Type" parameter:

https://lookups.twilio.com/v1/PhoneNumbers/6156?Type=carrier&Type=caller-name

But because the twilify() method expects either a hash or something
that can be reasonably stringfied, it is impossible to specify "Type" more
than once.

This change allows one to provide an array of [name,value] pairs as
a parameter list to the phone number lookup's "get", by modifying
twilify() to understand lists of pairs.