Avoids issues with ActionController::Parameters.
From Rails 5.1 the Hash methods on ActionController::Parameters will be removed, from Rails 5 they are deprecated. This is because ActionController::Parameters no longer inherits from Hash (well, ActiveSupport::HashWithIndifferentAccess, but who's counting?).

Twilio::Util::RequestValidator users Hash#sort, so we need to convert an ActionController::Parameters object to a hash so that we can continue to validate requests in this way.