Fixes use of configuration in client.
As reported in #227, when you rely on the config set in the Twilio.configure block, the client tries to use the private class method configuration to access the account_sid and auth_token. There are direct readers for this on the Twilio module which we should use instead.

I've made this a pull request just to highlight the change though I understand that this is part of the generated code.