Fixes next gen list resource pagination.
When paginating in a next gen list resource using next_page subsequent instance resources were initialized with a full URL including the pagination parameters for a path. Trying to carry out any operations on that instance would then fail as the URL was really a list url.

This commit fixes this by splitting the pagination URL up and returning the list resource path when it is a full_url.