Making message and code accessible when response is a kind of Net::HTTPBadRequest
Net::HTTPClientError is the parent of Net::HTTPBadRequest.

But when the response is a Net::HTTPBadRequest, the object hash uses symbols as keys, not strings.  So the Twilio::REST::RequestError created in line 125 would not have the message or the code values.

I don't know how object will be used when returned by this method (laziness!), so making object an instance of HashWithIndifferentAccess when set in line 121, so that either type of key can be used.