Making message and code accessible when response is a kind of Net::HTTPBadRequest
Net::HTTPClientError is the parent of Net::HTTPBadRequest.

But when the response is a Net::HTTPBadRequest, the object hash uses symbols as keys, not strings. So the Twilio::REST::RequestError created in line 125 would not have the message or the code values.

The hash returned by connect_and_send is not used by the caller so using string keys when it is set in line 121 enables these keys to be used in line 125.