Support API credentials in Twilio.configure
[Twilio docs state](https://www.twilio.com/docs/api/rest/request#credentials)
that using AccountSid and AuthToken can be replaced with API credentials to
authenticate into the platform. Some developers, including myself, thought that
the API keys could be drop-in replacements for the Twilio.configure block,
which is untrue.

This maintains backwards compatibility but allows for developers to universally
configure the Twilio client to use their API credentials for authentication
and set the appropriate AccountSid context for Message creation and other calls