Uses Nokogiri to better generate XML.
This leads to XML that has line breaks after each element, which wasn't present before, but doesn't break anything in terms of the expected output. This follows #382, which updated the tests to not worry about insignificant whitespace. The purpose of this PR is to properly use all of Nokogiri's features to generate the XML, including the header, and fixes what seems like an inconsistency, in which the XML is generated from a  instead of the whole .