Remove jruby-openssl requirement. (#436)
Turns out we didn't need jruby-openssl. I thought it was required for
the signature verification but it turns out that JRuby ships enough
OpenSSL to do what we need (https://www.jruby.org/openssl).

Forcing an install of bundler on Travis is unnecessary too. Travis will
install a version of bundler and as long as it is within our supported
range it should work. Forcing an install will make Travis try to install
bundler on Ruby 2.2 which no longer works as Bundler 2 only supports
Ruby 2.3 and up.