fix: don't load webhook authentication if Rack not present
Now that the webhook authentication requires part of Rack to run, loading the gem in a context that doesn't include Rack fails. For most people this won't happen as it will be used in the context of a web framework and most Ruby web frameworks rely on Rack. However, simple scripts that don't have Rack loaded will fail because the application can't load rack/media_type. In these cases, since Rack is not present, rack middleware won't be helpful either, so we only load the Rack middleware if Rack is present.