chore: `bundle install` with `--with` option is dispensable in this case
The setup script and the ci workflow definition specify that bundler should use
`--with development` option, which has been deprecated as of newer version of bundler.

While the intention apparently is to restrict installed gems to development group,
it makes no difference to the resulting installation set.

This is because we cannot rely on `--with` to opt out from installing grouped gems.
What we might want to use is either `--without` option or `group :test, optional: true`.
A relevant reference is here:
https://bundler.io/guides/groups.html#optional-groups-and-bundlewith

The discussion above in turn shows that `gem 'simplecov'` does not need to
reside in test group, as it's been, and will be, installed regardless of
the presence of `--with`. Therefore it's got rid of as well.